phase-3/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary>=2.9.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
better-exceptions==0.3.2
httpx==0.25.2
pydantic-settings==2.6.1
openai>=1.10.0